requirements.txt
discord.py
nltk
joblib
